requirements.txt
pip==9.0.1
Flask==0.10.1
EmailHarvester==2.0.1
argparse==1.4.0
flask-restplus==0.9.2
pycrypto==2.6.1
